Abstracted, renamed and moved components
