+++ src/components/Collection/Episode/EpisodeSummary.tsx
@@ -19,11 +19,10 @@ import EpisodeFiles from './EpisodeFiles';
 import type { EpisodeType } from '@/core/types/api/episode';
 
 type Props = {
-  anidbSeriesId?: number;
+  animeId?: number;
   episode: EpisodeType;
   nextUp?: boolean;
   page?: number;
-  seriesId: number;
 };
 
 const StateIcon = ({ icon, show }: { icon: string, show: boolean }) => (
@@ -44,7 +43,7 @@ const StateButton = React.memo((
   </Button>
 ));
 
-const SeriesEpisode = React.memo(({ anidbSeriesId, episode, nextUp, page, seriesId }: Props) => {
+const EpisodeSummary = React.memo(({ animeId, episode, nextUp, page }: Props) => {
   const thumbnail = useEpisodeThumbnail(episode);
   const [open, toggleOpen] = useToggle(false);
   const episodeId = get(episode, 'IDs.ID', 0);
@@ -54,8 +53,8 @@ const SeriesEpisode = React.memo(({ anidbSeriesId, episode, nextUp, page, series
     { includeDataFrom: ['AniDB'], include: ['AbsolutePaths', 'MediaInfo'] },
     open,
   );
-  const { mutate: markWatched } = useWatchEpisodeMutation(seriesId, page, nextUp);
-  const { mutate: markHidden } = useHideEpisodeMutation(seriesId, nextUp);
+  const { mutate: markWatched } = useWatchEpisodeMutation(page, nextUp);
+  const { mutate: markHidden } = useHideEpisodeMutation(nextUp);
 
   const handleMarkWatched = useEventCallback(() => markWatched({ episodeId, watched: episode.Watched === null }));
   const handleMarkHidden = useEventCallback(() => markHidden({ episodeId, hidden: !episode.IsHidden }));
@@ -96,7 +95,7 @@ const SeriesEpisode = React.memo(({ anidbSeriesId, episode, nextUp, page, series
         </BackgroundImagePlaceholderDiv>
         <EpisodeDetails episode={episode} />
       </div>
-      {!nextUp && anidbSeriesId && episode.Size > 0 && (
+      {animeId && episode.Size > 0 && (
         <>
           <div
             className="flex cursor-pointer justify-center gap-x-4 border-t-2 border-panel-border py-4 font-semibold"
@@ -112,11 +111,7 @@ const SeriesEpisode = React.memo(({ anidbSeriesId, episode, nextUp, page, series
             />
           </div>
           <AnimateHeight height={open && episodeFilesQuery.isSuccess ? 'auto' : 0}>
-            <EpisodeFiles
-              anidbSeriesId={anidbSeriesId}
-              episodeFiles={episodeFilesQuery.data ?? []}
-              episodeId={episodeId}
-            />
+            <EpisodeFiles animeId={animeId} episodeFiles={episodeFilesQuery.data ?? []} episodeId={episodeId} />
           </AnimateHeight>
         </>
       )}
@@ -124,4 +119,4 @@ const SeriesEpisode = React.memo(({ anidbSeriesId, episode, nextUp, page, series
   );
 });
 
-export default SeriesEpisode;
+export default EpisodeSummary;
